remove `to` and `from` from `MatrixLike`
These concepts were pretty confusing, as it represents a `Point` on the `Matrix` class but a point on the _parent_ of a `Lens`. This inconsistency made it a lot harder to work with these classes.

Instead, we can just avoid making this public state at all. The few places where it was necessary can be changed to use the height and width instead, which are still public.

This had a side-effect of resolving a bug in the `Inset` class, where some coordinates were not being computed correctly, likely because of the confusion around how and when to use the `to` and `from` properties correctly.