improve logic w/r/t writing empty cells between frames
Previously, the rendering system worked with the assumption that each Drawable would be a single instance throughout the lifetime of the application. Drawables needed to keep track of where they needed to write an empty cells if their content changed in a way where they were no longer writing to a location that needed to be written to

Now that we create new Drawables one each render, the Backend can/should/now does keep track of what the previous state was before rendering, determines where it needs to write empty cells, and takes care of that for us!